Pull out DummyManifestPlugin so it can be reused
